6.8 - Created AnecdoteList.js
